% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filmpalettes.R
\name{film_palette}
\alias{film_palette}
\title{A famous film palette generator}
\usage{
film_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette.}

\item{n}{Number of colors desired. The maximum number is 10. All colour
schemes are derived from this Esquire article:
\href{http://www.esquire.com/style/news/g2760/color-palettes-iconic-films/}{50 Iconic Films and Their Color Palettes}.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of colour palettes from famous films movies. An expansion to the \href{https://github.com/karthik/wesanderson}{Wes Anderson palette generator}.
}
\examples{
film_palette("TheMatrix")
film_palette("MadMax")
film_palette("Deadpool")

}
\keyword{colors}
