% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filmpalettes.R
\docType{data}
\name{film_palettes}
\alias{film_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 50.}
\usage{
film_palettes
}
\description{
Use \code{\link{film_palette}} to construct palettes of desired length.
}
\keyword{datasets}
